hoveredNodeIdをTreeAreaで管理するように修正
