+++ app/javascript/pages/trees/edit.tsx
@@ -18,7 +18,6 @@ const EditTreePage = () => {
   });
   const [selectedNodeIds, setSelectedNodeIds] = useState<number[]>([]);
   const [isLoading, setisLoading] = useState(true);
-  const [hoveredNodeId, setHoveredNodeId] = useState<number | null>(null);
 
   useEffect(() => {
     const load = async () => {
@@ -34,7 +33,7 @@ const EditTreePage = () => {
       }
     };
     load();
-  }, []);
+  }, [treeId]);
 
   if (isLoading) return <>ロード中…</>;
 
@@ -66,20 +65,6 @@ const EditTreePage = () => {
     setSelectedNodeIds(selectedNodeIds);
   };
 
-  const handleMouseOver: TreeNodeEventCallback = (node) => {
-    const hoveredNode = treeData.nodes.find(
-      (nodeData) => nodeData.id === node.data?.attributes?.id
-    );
-    if (!hoveredNode) {
-      return;
-    }
-    setHoveredNodeId(hoveredNode.id);
-  };
-
-  const handleMouseOut: TreeNodeEventCallback = () => {
-    setHoveredNodeId(null);
-  };
-
   return (
     <>
       <div className="flex w-full">
@@ -96,9 +81,6 @@ const EditTreePage = () => {
               treeData={treeData}
               selectedNodeIds={selectedNodeIds}
               handleClick={handleClick}
-              hoveredNodeId={hoveredNodeId}
-              handleMouseOver={handleMouseOver}
-              handleMouseOut={handleMouseOut}
               onUpdateSuccess={handleUpdateSuccess}
             />
           </ErrorBoundary>
