+++ spec/javascript/components/trees/tree/TreeArea.spec.tsx
@@ -14,9 +14,6 @@ describe.skip("ツリーの表示", () => {
         treeData: fixtures.treeData,
         selectedNodeIds: [],
         handleClick,
-        hoveredNodeId: null,
-        handleMouseOver: jest.fn(),
-        handleMouseOut: jest.fn(),
         onUpdateSuccess: jest.fn(),
       };
       const { container } = render(<TreeArea {...props} />);
